app/qa_service/requirements.txt
fastapi==0.103.2
langchain==0.0.316
weaviate_client==3.24.2
uvicorn==0.23.2
sentence-transformers
openai

